+++ lib/serial/serial_none.dart
